backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
numpy
Pillow
scikit-learn
h5py
torch
diffusers
transformers
accelerate
controlnet_aux
opencv-python
python-dotenv
requests
